// packages/twenty-apps/internal/gsh-events/src/types/service-order-update.ts
export type ServiceOrderUpdate = {
  enabled: boolean;
  id?: string;
  name?: string;
  status?: string;
  timeline?: string;
  finalMenu?: string;
  structureAndEquipment?: string;
  teamGuidance?: string;
  responsible?: string;
  distributionStatus?: string;
  distributedAt?: string;
  documentUrl?: string;
};